Moved auth.service into is own folder Created favorites.service
