+++ src/app/services/auth/auth.service.ts
@@ -3,9 +3,9 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Subject } from 'rxjs/Subject';
 
-import { User } from '../models/auth/user';
+import { User } from '../../models/auth/user';
 
-import * as App from '../app.config';
+import * as App from '../../app.config';
 import * as _ from 'lodash';
 import { Router } from '@angular/router';
