+++ src/app/services/services.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AuthService } from './auth.service';
+import { AuthService } from './auth/auth.service';
 import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
